Revert "GEODE-6425: Fixes rollover in some cache stats (#3214)"
This reverts commit c00e1c79ae8fdf4d46e78958275bdde840d81538.

+++ geode-core/src/integrationTest/java/org/apache/geode/TXJUnitTest.java
@@ -4832,10 +4832,10 @@ public class TXJUnitTest {
   public void testCacheStats() throws CacheException {
     CachePerfStats cacheStats = this.cache.getCachePerfStats();
     // quick sanity check to make sure perf stats work non-tx
-    long creates;
-    long destroys;
-    long puts;
-    long invalidates;
+    int creates;
+    int destroys;
+    int puts;
+    int invalidates;
 
     creates = cacheStats.getCreates();
     destroys = cacheStats.getDestroys();
